move parameter filter utils to own file (#18634)
